pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "df-lab-cli"
version = "0.1.0"
description = "CLI tool for Dwarf Fortress–style unit generation and evaluation"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "David Vidal" }]
dependencies = ["PyYAML>=6.0.2"]

[project.optional-dependencies]
dev = ["pytest", "python-lsp-server[all]", "black", "flake8"]

[project.scripts]
dflab = "dflab_cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dflab_cli*"]
